Numbat R6 class

Used to allow users to plot results

Public fields

label

character Sample name

gtf

dataframe Transcript annotation

joint_post

dataframe Joint posterior

exp_post

dataframe Expression posterior

allele_post

dataframe Allele posetrior

bulk_subtrees

dataframe Bulk profiles of lineage subtrees

bulk_clones

dataframe Bulk profiles of clones

segs_consensus

dataframe Consensus segments

tree_post

list Tree posterior

mut_graph

igraph Mutation history graph

gtree

tbl_graph Single-cell phylogeny

clone_post

dataframe Clone posteriors

gexp_roll_wide

matrix Smoothed expression of single cells

hc

hclust Initial hierarchical clustering

Methods

Public methods

Numbat$new()

Numbat$plot_phylo_heatmap()

Numbat$plot_exp_roll()

Numbat$plot_mut_history()

Numbat$plot_sc_tree()

Numbat$plot_consensus()

Numbat$plot_clone_profile()

Numbat$clone()

Method new()

initialize Numbat class

Usage

Numbat$new(out_dir, i = 2, gtf = gtf_hg38, verbose = TRUE)

Arguments

out_dir

character string Output directory

i

integer Get results from which iteration (either 1 or 2) (default=2)

gtf

dataframe Transcript gtf (default=gtf_hg38)

verbose

logical Whether to output verbose results (default=TRUE)

Returns

a new 'Numbat' object

Method plot_phylo_heatmap()

Plot the single-cell CNV calls in a heatmap and the corresponding phylogeny

Usage

Numbat$plot_phylo_heatmap(...)

Arguments

...

additional parameters passed to plot_phylo_heatmap()

Method plot_exp_roll()

Plot window-smoothed expression profiles

Usage

Numbat$plot_exp_roll(k = 3, n_sample = 300, ...)

Arguments

k

integer Number of clusters

n_sample

integer Number of cells to subsample

...

additional parameters passed to plot_exp_roll()

Method plot_mut_history()

Plot the mutation history of the tumor

Usage

Numbat$plot_mut_history(...)

Arguments

...

additional parameters passed to plot_mut_history()

Method plot_sc_tree()

Plot the single cell phylogeny

Usage

Numbat$plot_sc_tree(...)

Arguments

...

additional parameters passed to plot_sc_tree()

Method plot_consensus()

Plot consensus segments

Usage

Numbat$plot_consensus(...)

Arguments

...

additional parameters passed to plot_sc_tree()

Method plot_clone_profile()

Plot clone cnv profiles

Usage

Numbat$plot_clone_profile(...)

Arguments

...

additional parameters passed to plot_clone_profile()

Method clone()

The objects of this class are cloneable with this method.

Usage

Numbat$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.