Beta-binomial distribution density function A distribution is beta-binomial if p, the probability of success, in a binomial distribution has a beta distribution with shape parameters alpha > 0 and beta > 0 For more details, see extraDistr::dbbinom

Beta-binomial distribution density function A distribution is beta-binomial if p, the probability of success, in a binomial distribution has a beta distribution with shape parameters alpha > 0 and beta > 0 For more details, see extraDistr::dbbinom

Usage

dbbinom(x, size, alpha = 1, beta = 1, log = FALSE)

Arguments

x

vector of quantiles

size

number of trials (zero or more)

alpha

numeric (default=1)

beta

numeric (default=1)

log

boolean (default=FALSE)

Value

density values returned as numeric vector

Examples

xx <- 1:1000
dbbinom(xx, 1000, 5, 13)
#>    [1] 3.405595e-09 1.009552e-08 2.327633e-08 4.599901e-08 8.181252e-08
#>    [6] 1.347293e-07 2.091921e-07 3.100414e-07 4.424849e-07 6.120673e-07
#>   [11] 8.246416e-07 1.086341e-06 1.403553e-06 1.782891e-06 2.231175e-06
#>   [16] 2.755400e-06 3.362721e-06 4.060424e-06 4.855911e-06 5.756675e-06
#>   [21] 6.770283e-06 7.904357e-06 9.166555e-06 1.056455e-05 1.210604e-05
#>   [26] 1.379868e-05 1.565011e-05 1.766794e-05 1.985972e-05 2.223292e-05
#>   [31] 2.479494e-05 2.755309e-05 3.051458e-05 3.368650e-05 3.707581e-05
#>   [36] 4.068937e-05 4.453385e-05 4.861582e-05 5.294167e-05 5.751761e-05
#>   [41] 6.234971e-05 6.744386e-05 7.280573e-05 7.844085e-05 8.435453e-05
#>   [46] 9.055188e-05 9.703782e-05 1.038171e-04 1.108941e-04 1.182732e-04
#>   [51] 1.259585e-04 1.339538e-04 1.422627e-04 1.508887e-04 1.598349e-04
#>   [56] 1.691043e-04 1.786997e-04 1.886235e-04 1.988781e-04 2.094654e-04
#>   [61] 2.203874e-04 2.316456e-04 2.432414e-04 2.551760e-04 2.674503e-04
#>   [66] 2.800650e-04 2.930206e-04 3.063173e-04 3.199552e-04 3.339342e-04
#>   [71] 3.482538e-04 3.629134e-04 3.779123e-04 3.932494e-04 4.089235e-04
#>   [76] 4.249331e-04 4.412767e-04 4.579524e-04 4.749582e-04 4.922919e-04
#>   [81] 5.099510e-04 5.279331e-04 5.462354e-04 5.648548e-04 5.837883e-04
#>   [86] 6.030326e-04 6.225843e-04 6.424397e-04 6.625950e-04 6.830464e-04
#>   [91] 7.037897e-04 7.248206e-04 7.461349e-04 7.677279e-04 7.895951e-04
#>   [96] 8.117316e-04 8.341325e-04 8.567928e-04 8.797073e-04 9.028706e-04
#>  [101] 9.262775e-04 9.499224e-04 9.737997e-04 9.979036e-04 1.022228e-03
#>  [106] 1.046768e-03 1.071517e-03 1.096469e-03 1.121617e-03 1.146956e-03
#>  [111] 1.172479e-03 1.198180e-03 1.224052e-03 1.250089e-03 1.276284e-03
#>  [116] 1.302631e-03 1.329123e-03 1.355754e-03 1.382515e-03 1.409402e-03
#>  [121] 1.436406e-03 1.463522e-03 1.490741e-03 1.518058e-03 1.545465e-03
#>  [126] 1.572956e-03 1.600523e-03 1.628159e-03 1.655857e-03 1.683611e-03
#>  [131] 1.711413e-03 1.739257e-03 1.767135e-03 1.795041e-03 1.822966e-03
#>  [136] 1.850906e-03 1.878852e-03 1.906797e-03 1.934735e-03 1.962659e-03
#>  [141] 1.990562e-03 2.018438e-03 2.046278e-03 2.074077e-03 2.101828e-03
#>  [146] 2.129525e-03 2.157159e-03 2.184726e-03 2.212219e-03 2.239630e-03
#>  [151] 2.266954e-03 2.294184e-03 2.321314e-03 2.348337e-03 2.375248e-03
#>  [156] 2.402040e-03 2.428707e-03 2.455243e-03 2.481643e-03 2.507899e-03
#>  [161] 2.534007e-03 2.559961e-03 2.585754e-03 2.611382e-03 2.636839e-03
#>  [166] 2.662119e-03 2.687218e-03 2.712129e-03 2.736848e-03 2.761369e-03
#>  [171] 2.785687e-03 2.809798e-03 2.833696e-03 2.857377e-03 2.880836e-03
#>  [176] 2.904069e-03 2.927070e-03 2.949836e-03 2.972362e-03 2.994644e-03
#>  [181] 3.016678e-03 3.038459e-03 3.059983e-03 3.081248e-03 3.102248e-03
#>  [186] 3.122980e-03 3.143442e-03 3.163628e-03 3.183535e-03 3.203161e-03
#>  [191] 3.222503e-03 3.241556e-03 3.260318e-03 3.278785e-03 3.296956e-03
#>  [196] 3.314828e-03 3.332397e-03 3.349661e-03 3.366617e-03 3.383264e-03
#>  [201] 3.399598e-03 3.415619e-03 3.431323e-03 3.446708e-03 3.461774e-03
#>  [206] 3.476517e-03 3.490936e-03 3.505030e-03 3.518797e-03 3.532235e-03
#>  [211] 3.545344e-03 3.558121e-03 3.570566e-03 3.582678e-03 3.594455e-03
#>  [216] 3.605897e-03 3.617003e-03 3.627772e-03 3.638203e-03 3.648297e-03
#>  [221] 3.658052e-03 3.667467e-03 3.676544e-03 3.685281e-03 3.693678e-03
#>  [226] 3.701736e-03 3.709454e-03 3.716832e-03 3.723871e-03 3.730571e-03
#>  [231] 3.736931e-03 3.742954e-03 3.748638e-03 3.753985e-03 3.758995e-03
#>  [236] 3.763668e-03 3.768007e-03 3.772011e-03 3.775681e-03 3.779019e-03
#>  [241] 3.782025e-03 3.784701e-03 3.787047e-03 3.789065e-03 3.790757e-03
#>  [246] 3.792123e-03 3.793165e-03 3.793885e-03 3.794284e-03 3.794364e-03
#>  [251] 3.794125e-03 3.793571e-03 3.792703e-03 3.791523e-03 3.790032e-03
#>  [256] 3.788233e-03 3.786127e-03 3.783717e-03 3.781004e-03 3.777992e-03
#>  [261] 3.774681e-03 3.771074e-03 3.767174e-03 3.762983e-03 3.758503e-03
#>  [266] 3.753736e-03 3.748685e-03 3.743353e-03 3.737742e-03 3.731854e-03
#>  [271] 3.725693e-03 3.719260e-03 3.712559e-03 3.705591e-03 3.698361e-03
#>  [276] 3.690870e-03 3.683122e-03 3.675118e-03 3.666863e-03 3.658359e-03
#>  [281] 3.649608e-03 3.640615e-03 3.631380e-03 3.621909e-03 3.612203e-03
#>  [286] 3.602266e-03 3.592100e-03 3.581709e-03 3.571096e-03 3.560264e-03
#>  [291] 3.549215e-03 3.537954e-03 3.526483e-03 3.514805e-03 3.502923e-03
#>  [296] 3.490842e-03 3.478563e-03 3.466090e-03 3.453426e-03 3.440574e-03
#>  [301] 3.427538e-03 3.414321e-03 3.400926e-03 3.387356e-03 3.373615e-03
#>  [306] 3.359705e-03 3.345630e-03 3.331394e-03 3.316998e-03 3.302447e-03
#>  [311] 3.287744e-03 3.272893e-03 3.257895e-03 3.242755e-03 3.227475e-03
#>  [316] 3.212060e-03 3.196511e-03 3.180833e-03 3.165029e-03 3.149101e-03
#>  [321] 3.133053e-03 3.116888e-03 3.100609e-03 3.084219e-03 3.067722e-03
#>  [326] 3.051121e-03 3.034418e-03 3.017618e-03 3.000722e-03 2.983734e-03
#>  [331] 2.966657e-03 2.949494e-03 2.932248e-03 2.914922e-03 2.897520e-03
#>  [336] 2.880043e-03 2.862496e-03 2.844881e-03 2.827200e-03 2.809458e-03
#>  [341] 2.791656e-03 2.773798e-03 2.755886e-03 2.737923e-03 2.719913e-03
#>  [346] 2.701857e-03 2.683759e-03 2.665622e-03 2.647447e-03 2.629238e-03
#>  [351] 2.610998e-03 2.592729e-03 2.574434e-03 2.556115e-03 2.537775e-03
#>  [356] 2.519416e-03 2.501042e-03 2.482654e-03 2.464255e-03 2.445847e-03
#>  [361] 2.427434e-03 2.409016e-03 2.390598e-03 2.372180e-03 2.353766e-03
#>  [366] 2.335358e-03 2.316957e-03 2.298567e-03 2.280189e-03 2.261825e-03
#>  [371] 2.243479e-03 2.225151e-03 2.206844e-03 2.188560e-03 2.170302e-03
#>  [376] 2.152070e-03 2.133868e-03 2.115697e-03 2.097559e-03 2.079456e-03
#>  [381] 2.061389e-03 2.043362e-03 2.025375e-03 2.007430e-03 1.989529e-03
#>  [386] 1.971674e-03 1.953867e-03 1.936109e-03 1.918402e-03 1.900747e-03
#>  [391] 1.883147e-03 1.865602e-03 1.848114e-03 1.830685e-03 1.813317e-03
#>  [396] 1.796010e-03 1.778766e-03 1.761586e-03 1.744473e-03 1.727427e-03
#>  [401] 1.710449e-03 1.693541e-03 1.676704e-03 1.659939e-03 1.643248e-03
#>  [406] 1.626632e-03 1.610091e-03 1.593628e-03 1.577242e-03 1.560936e-03
#>  [411] 1.544710e-03 1.528565e-03 1.512502e-03 1.496522e-03 1.480626e-03
#>  [416] 1.464815e-03 1.449091e-03 1.433453e-03 1.417902e-03 1.402440e-03
#>  [421] 1.387067e-03 1.371783e-03 1.356591e-03 1.341490e-03 1.326480e-03
#>  [426] 1.311564e-03 1.296741e-03 1.282011e-03 1.267376e-03 1.252837e-03
#>  [431] 1.238393e-03 1.224044e-03 1.209793e-03 1.195639e-03 1.181582e-03
#>  [436] 1.167623e-03 1.153762e-03 1.140001e-03 1.126338e-03 1.112775e-03
#>  [441] 1.099311e-03 1.085948e-03 1.072685e-03 1.059522e-03 1.046461e-03
#>  [446] 1.033500e-03 1.020641e-03 1.007883e-03 9.952261e-04 9.826713e-04
#>  [451] 9.702184e-04 9.578674e-04 9.456185e-04 9.334715e-04 9.214268e-04
#>  [456] 9.094841e-04 8.976436e-04 8.859053e-04 8.742691e-04 8.627350e-04
#>  [461] 8.513029e-04 8.399728e-04 8.287446e-04 8.176181e-04 8.065933e-04
#>  [466] 7.956701e-04 7.848482e-04 7.741275e-04 7.635079e-04 7.529892e-04
#>  [471] 7.425710e-04 7.322534e-04 7.220359e-04 7.119183e-04 7.019005e-04
#>  [476] 6.919821e-04 6.821628e-04 6.724424e-04 6.628205e-04 6.532969e-04
#>  [481] 6.438712e-04 6.345430e-04 6.253120e-04 6.161779e-04 6.071402e-04
#>  [486] 5.981986e-04 5.893528e-04 5.806022e-04 5.719465e-04 5.633853e-04
#>  [491] 5.549181e-04 5.465445e-04 5.382640e-04 5.300762e-04 5.219807e-04
#>  [496] 5.139769e-04 5.060644e-04 4.982426e-04 4.905112e-04 4.828696e-04
#>  [501] 4.753172e-04 4.678536e-04 4.604782e-04 4.531906e-04 4.459901e-04
#>  [506] 4.388763e-04 4.318486e-04 4.249064e-04 4.180493e-04 4.112765e-04
#>  [511] 4.045876e-04 3.979820e-04 3.914592e-04 3.850184e-04 3.786593e-04
#>  [516] 3.723811e-04 3.661832e-04 3.600652e-04 3.540263e-04 3.480660e-04
#>  [521] 3.421837e-04 3.363788e-04 3.306506e-04 3.249986e-04 3.194221e-04
#>  [526] 3.139206e-04 3.084933e-04 3.031397e-04 2.978592e-04 2.926511e-04
#>  [531] 2.875149e-04 2.824498e-04 2.774552e-04 2.725306e-04 2.676753e-04
#>  [536] 2.628887e-04 2.581700e-04 2.535188e-04 2.489344e-04 2.444161e-04
#>  [541] 2.399634e-04 2.355755e-04 2.312518e-04 2.269918e-04 2.227948e-04
#>  [546] 2.186601e-04 2.145872e-04 2.105754e-04 2.066240e-04 2.027325e-04
#>  [551] 1.989003e-04 1.951266e-04 1.914109e-04 1.877526e-04 1.841511e-04
#>  [556] 1.806056e-04 1.771157e-04 1.736807e-04 1.702999e-04 1.669729e-04
#>  [561] 1.636989e-04 1.604774e-04 1.573078e-04 1.541894e-04 1.511217e-04
#>  [566] 1.481041e-04 1.451359e-04 1.422166e-04 1.393457e-04 1.365225e-04
#>  [571] 1.337464e-04 1.310169e-04 1.283333e-04 1.256952e-04 1.231020e-04
#>  [576] 1.205530e-04 1.180477e-04 1.155857e-04 1.131662e-04 1.107888e-04
#>  [581] 1.084529e-04 1.061579e-04 1.039034e-04 1.016888e-04 9.951350e-05
#>  [586] 9.737705e-05 9.527890e-05 9.321853e-05 9.119541e-05 8.920904e-05
#>  [591] 8.725891e-05 8.534450e-05 8.346532e-05 8.162086e-05 7.981064e-05
#>  [596] 7.803416e-05 7.629093e-05 7.458048e-05 7.290233e-05 7.125599e-05
#>  [601] 6.964101e-05 6.805692e-05 6.650325e-05 6.497955e-05 6.348537e-05
#>  [606] 6.202025e-05 6.058376e-05 5.917546e-05 5.779490e-05 5.644166e-05
#>  [611] 5.511530e-05 5.381542e-05 5.254158e-05 5.129338e-05 5.007040e-05
#>  [616] 4.887225e-05 4.769851e-05 4.654879e-05 4.542269e-05 4.431984e-05
#>  [621] 4.323985e-05 4.218233e-05 4.114691e-05 4.013323e-05 3.914090e-05
#>  [626] 3.816958e-05 3.721889e-05 3.628850e-05 3.537804e-05 3.448717e-05
#>  [631] 3.361556e-05 3.276286e-05 3.192873e-05 3.111286e-05 3.031492e-05
#>  [636] 2.953458e-05 2.877153e-05 2.802545e-05 2.729604e-05 2.658300e-05
#>  [641] 2.588602e-05 2.520480e-05 2.453906e-05 2.388850e-05 2.325284e-05
#>  [646] 2.263180e-05 2.202511e-05 2.143248e-05 2.085366e-05 2.028837e-05
#>  [651] 1.973635e-05 1.919735e-05 1.867112e-05 1.815739e-05 1.765593e-05
#>  [656] 1.716649e-05 1.668884e-05 1.622273e-05 1.576794e-05 1.532423e-05
#>  [661] 1.489139e-05 1.446918e-05 1.405740e-05 1.365582e-05 1.326424e-05
#>  [666] 1.288244e-05 1.251023e-05 1.214740e-05 1.179375e-05 1.144909e-05
#>  [671] 1.111322e-05 1.078596e-05 1.046712e-05 1.015653e-05 9.853990e-06
#>  [676] 9.559338e-06 9.272396e-06 8.992996e-06 8.720969e-06 8.456151e-06
#>  [681] 8.198380e-06 7.947499e-06 7.703351e-06 7.465784e-06 7.234646e-06
#>  [686] 7.009791e-06 6.791074e-06 6.578351e-06 6.371485e-06 6.170337e-06
#>  [691] 5.974774e-06 5.784663e-06 5.599875e-06 5.420284e-06 5.245763e-06
#>  [696] 5.076192e-06 4.911451e-06 4.751422e-06 4.595989e-06 4.445041e-06
#>  [701] 4.298467e-06 4.156157e-06 4.018006e-06 3.883910e-06 3.753767e-06
#>  [706] 3.627476e-06 3.504941e-06 3.386064e-06 3.270753e-06 3.158915e-06
#>  [711] 3.050461e-06 2.945302e-06 2.843352e-06 2.744528e-06 2.648746e-06
#>  [716] 2.555925e-06 2.465987e-06 2.378855e-06 2.294453e-06 2.212707e-06
#>  [721] 2.133546e-06 2.056897e-06 1.982693e-06 1.910866e-06 1.841350e-06
#>  [726] 1.774081e-06 1.708995e-06 1.646032e-06 1.585132e-06 1.526235e-06
#>  [731] 1.469285e-06 1.414226e-06 1.361003e-06 1.309563e-06 1.259854e-06
#>  [736] 1.211826e-06 1.165429e-06 1.120615e-06 1.077337e-06 1.035549e-06
#>  [741] 9.952064e-07 9.562656e-07 9.186843e-07 8.824209e-07 8.474352e-07
#>  [746] 8.136879e-07 7.811407e-07 7.497564e-07 7.194986e-07 6.903320e-07
#>  [751] 6.622222e-07 6.351357e-07 6.090399e-07 5.839031e-07 5.596945e-07
#>  [756] 5.363840e-07 5.139424e-07 4.923413e-07 4.715532e-07 4.515512e-07
#>  [761] 4.323092e-07 4.138019e-07 3.960046e-07 3.788934e-07 3.624452e-07
#>  [766] 3.466372e-07 3.314476e-07 3.168552e-07 3.028392e-07 2.893797e-07
#>  [771] 2.764572e-07 2.640528e-07 2.521482e-07 2.407257e-07 2.297681e-07
#>  [776] 2.192587e-07 2.091813e-07 1.995202e-07 1.902604e-07 1.813870e-07
#>  [781] 1.728859e-07 1.647432e-07 1.569456e-07 1.494801e-07 1.423344e-07
#>  [786] 1.354961e-07 1.289537e-07 1.226959e-07 1.167116e-07 1.109903e-07
#>  [791] 1.055218e-07 1.002961e-07 9.530368e-08 9.053538e-08 8.598223e-08
#>  [796] 8.163563e-08 7.748727e-08 7.352913e-08 6.975346e-08 6.615281e-08
#>  [801] 6.271996e-08 5.944798e-08 5.633016e-08 5.336005e-08 5.053143e-08
#>  [806] 4.783831e-08 4.527492e-08 4.283569e-08 4.051528e-08 3.830854e-08
#>  [811] 3.621050e-08 3.421641e-08 3.232167e-08 3.052188e-08 2.881279e-08
#>  [816] 2.719033e-08 2.565059e-08 2.418980e-08 2.280436e-08 2.149080e-08
#>  [821] 2.024578e-08 1.906613e-08 1.794877e-08 1.689076e-08 1.588929e-08
#>  [826] 1.494167e-08 1.404529e-08 1.319770e-08 1.239651e-08 1.163945e-08
#>  [831] 1.092434e-08 1.024912e-08 9.611775e-09 9.010419e-09 8.443229e-09
#>  [836] 7.908466e-09 7.404470e-09 6.929655e-09 6.482507e-09 6.061581e-09
#>  [841] 5.665499e-09 5.292946e-09 4.942668e-09 4.613471e-09 4.304217e-09
#>  [846] 4.013820e-09 3.741250e-09 3.485523e-09 3.245705e-09 3.020907e-09
#>  [851] 2.810283e-09 2.613032e-09 2.428389e-09 2.255630e-09 2.094067e-09
#>  [856] 1.943049e-09 1.801955e-09 1.670199e-09 1.547225e-09 1.432506e-09
#>  [861] 1.325543e-09 1.225864e-09 1.133022e-09 1.046595e-09 9.661834e-10
#>  [866] 8.914097e-10 8.219178e-10 7.573712e-10 6.974526e-10 6.418627e-10
#>  [871] 5.903194e-10 5.425568e-10 4.983248e-10 4.573877e-10 4.195240e-10
#>  [876] 3.845252e-10 3.521956e-10 3.223512e-10 2.948195e-10 2.694384e-10
#>  [881] 2.460561e-10 2.245303e-10 2.047277e-10 1.865234e-10 1.698009e-10
#>  [886] 1.544509e-10 1.403714e-10 1.274674e-10 1.156498e-10 1.048360e-10
#>  [891] 9.494861e-11 8.591578e-11 7.767056e-11 7.015071e-11 6.329837e-11
#>  [896] 5.705983e-11 5.138522e-11 4.622828e-11 4.154619e-11 3.729924e-11
#>  [901] 3.345074e-11 2.996675e-11 2.681591e-11 2.396929e-11 2.140021e-11
#>  [906] 1.908407e-11 1.699825e-11 1.512191e-11 1.343594e-11 1.192277e-11
#>  [911] 1.056628e-11 9.351719e-12 8.265567e-12 7.295469e-12 6.430134e-12
#>  [916] 5.659259e-12 4.973452e-12 4.364160e-12 3.823603e-12 3.344714e-12
#>  [921] 2.921079e-12 2.546883e-12 2.216865e-12 1.926264e-12 1.670786e-12
#>  [926] 1.446554e-12 1.250081e-12 1.078226e-12 9.281731e-13 7.973944e-13
#>  [931] 6.836271e-13 5.848484e-13 4.992524e-13 4.252300e-13 3.613493e-13
#>  [936] 3.063387e-13 2.590707e-13 2.185474e-13 1.838873e-13 1.543131e-13
#>  [941] 1.291409e-13 1.077699e-13 8.967385e-14 7.439229e-14 6.152356e-14
#>  [946] 5.071796e-14 4.167179e-14 3.412191e-14 2.784091e-14 2.263277e-14
#>  [951] 1.832901e-14 1.478516e-14 1.187778e-14 9.501630e-15 7.567339e-15
#>  [956] 5.999212e-15 4.733368e-15 3.716085e-15 2.902344e-15 2.254564e-15
#>  [961] 1.741499e-15 1.337272e-15 1.020548e-15 7.738155e-16 5.827673e-16
#>  [966] 4.357726e-16 3.234251e-16 2.381585e-16 1.739212e-16 1.259021e-16
#>  [971] 9.030053e-17 6.413395e-17 4.507832e-17 3.133623e-17 2.152854e-17
#>  [976] 1.460593e-17 9.777152e-18 6.451264e-18 4.191403e-18 2.678143e-18
#>  [981] 1.680664e-18 1.034280e-18 6.230935e-19 3.667465e-19 2.104205e-19
#>  [986] 1.173745e-19 6.345779e-20 3.313165e-20 1.663282e-20 7.986970e-21
#>  [991] 3.645094e-21 1.568482e-21 6.299202e-22 2.330098e-22 7.798217e-23
#>  [996] 2.302804e-23 5.780108e-24 1.160655e-24 1.664718e-25 1.285674e-26